fix(outbound): split delivery resolution types
