+++ crates/ironclaw_outbound/src/ids.rs
@@ -64,7 +64,7 @@ macro_rules! bounded_ref {
 
 bounded_ref!(ProjectionSubscriptionId, "projection_subscription_id");
 bounded_ref!(ProjectionUpdateRef, "projection_update_ref");
-bounded_ref!(TriggerId, "trigger_id");
+bounded_ref!(TriggerOriginRef, "trigger_origin_ref");
 bounded_ref!(TriggerFireSlot, "trigger_fire_slot");
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
@@ -135,7 +135,7 @@ mod tests {
     fn bounded_refs_reject_invalid_inputs() {
         assert_invalid_inputs!(ProjectionSubscriptionId, "projection_subscription_id");
         assert_invalid_inputs!(ProjectionUpdateRef, "projection_update_ref");
-        assert_invalid_inputs!(TriggerId, "trigger_id");
+        assert_invalid_inputs!(TriggerOriginRef, "trigger_origin_ref");
         assert_invalid_inputs!(TriggerFireSlot, "trigger_fire_slot");
     }
 }
